SIGHUP now reloads Haproxy config.
Defaults/globals/etc. service_config is currently not reloaded.

There is a race condition in Haproxy::State: if force_restart is
set and an updated state is received there will likely be two iterations of
the worker thread loop, leading to an unnecessary restart. We could
potentially be avoid this by checking the value of force_restart when an update
call is received, but I prefer to avoid complicating the logic.